requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.4
httpx==0.27.0
